ioctl commands are unsigned long.  ABI change to mrt6_ioctl() will affect 64-bit platforms.

+++ sys/netinet6/ip6_mroute.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: ip6_mroute.h,v 1.13 2007/11/01 20:33:57 dyoung Exp $	*/
+/*	$NetBSD: ip6_mroute.h,v 1.14 2008/06/24 10:35:15 gmcgarry Exp $	*/
 /*	$KAME: ip6_mroute.h,v 1.17 2001/02/10 02:05:52 itojun Exp $	*/
 
 /*
@@ -271,7 +271,7 @@ int	ip6_mrouter_set(int, struct socket *, struct mbuf *);
 int	ip6_mrouter_get(int, struct socket *, struct mbuf **);
 int	ip6_mrouter_done(void);
 void	ip6_mrouter_detach(struct ifnet *);
-int	mrt6_ioctl(int, void *);
+int	mrt6_ioctl(u_long, void *);
 #endif /* _KERNEL */
 
 #endif /* !_NETINET6_IP6_MROUTE_H_ */
